Feature/add verbose mode

+++ src/utils/working_directory.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import logging
 
 
 def create_and_get_into(path_to_working_directory):
@@ -12,6 +13,10 @@ def create_and_get_into(path_to_working_directory):
 
 
 def leave_and_delete(previous_directory):
-    working_directory = os.getcwd()
-    os.chdir(previous_directory)
-    shutil.rmtree(working_directory)
+    logger = logging.getLogger()
+    if logger.isEnabledFor(logging.DEBUG):
+        os.chdir(previous_directory)
+    else:
+        working_directory = os.getcwd()
+        os.chdir(previous_directory)
+        shutil.rmtree(working_directory)
